03/04/2025 Build
The encoding and display are basically working. I don't have any safeguards setup for the switch statements, so inputting random values won't cause the switch to loop. The display passenger information function is not yet fully functioning which is why it isn't in this build. Additionally, I have not yet paid any attention to the move passenger function or to the function that will check for priority